refactor(grpc_client): move main loop and other client init into util
